Added test case for updating a line chart from a container component, this catches the exception logged at https://github.com/recharts/recharts/issues/935
